--- app/NX/Shortcodes/components/GoogleMap.tsx
@@ -1,27 +1,0 @@
-// /Users/goldlabel/GitHub/core/gl-core/cartridges/Shortcodes/components/GoogleMap.tsx
-'use client';
-
-import * as React from 'react';
-import { Box } from '@mui/material';
-
-export type TGoogleMap = {
-  src?: string | null;
-};
-
-export default function GoogleMap({ src = null }: TGoogleMap) {
-  if (!src) return null;
-
-  return (
-    <Box>
-      <iframe
-        src={src}
-        style={{
-          border: 0,
-          borderRadius: '6px',
-        }}
-        width="100%"
-        height="380"
-      />
-    </Box>
-  );
-}
