+++ app/NX/Shortcodes/index.tsx
@@ -1,5 +1,9 @@
 import Shortcodes from "./Shortcodes";
+import BuyNow from "./components/BuyNow";
+import RenderMarkdown from "./components/RenderMarkdown";
 
 export {
     Shortcodes,
+    RenderMarkdown,
+    BuyNow,
 };
