+++ public/mcuk/markdown/index.md
@@ -10,5 +10,6 @@ tags: medical cannabis, uk, prescription, licensed pharmacies, CBPM, legal,
 
 #### Medical cannabis dispensed by licensed pharmacies. Delivered to your door
 
-> A comprehensive guide to medical cannabis in the UK, covering legal, medical, and practical aspects. Below is an overview of the available sections
+[BuyNow url="https://budbuddies.cc/product/ferrero-roche/bud/listingslab/?campaign=MCUK" label="Buy for £14.99"]
 
+> A comprehensive guide to medical cannabis in the UK, covering legal, medical, and practical aspects. Below is an overview of the available sections
